Move packet_commitment_bytes function
