+++ crates/relayer-solomachine/src/impls/chain/ofa.rs
@@ -55,7 +55,6 @@ use ibc_relayer_types::proofs::ConsensusProof;
 use ibc_relayer_types::timestamp::Timestamp;
 use ibc_relayer_types::tx_msg::Msg;
 use ibc_relayer_types::Height;
-use sha2::{Digest, Sha256};
 
 use crate::impls::chain::components::channel_handshake_payload::BuildSolomachineChannelHandshakePayloads;
 use crate::impls::chain::components::connection_handshake_payload::BuildSolomachineConnectionHandshakePayloads;
@@ -1157,19 +1156,4 @@ where
 
         Ok(message.to_cosmos_message())
     }
-}
-
-pub fn packet_commitment_bytes(packet: &Packet) -> Vec<u8> {
-    let mut buf = Vec::new();
-
-    let timeout_timestamp = packet.timeout_timestamp.nanoseconds();
-    let timeout_revision_number = packet.timeout_height.commitment_revision_number();
-    let timeout_revision_height = packet.timeout_height.commitment_revision_height();
-
-    buf.extend(timeout_timestamp.to_be_bytes());
-    buf.extend(timeout_revision_number.to_be_bytes());
-    buf.extend(timeout_revision_height.to_be_bytes());
-    buf.extend(Sha256::digest(&packet.data));
-
-    Sha256::digest(&buf).to_vec()
 }
