+++ crates/relayer-solomachine/src/methods/mod.rs
@@ -1,2 +1,3 @@
+pub mod commitment;
 pub mod encode;
 pub mod proofs;
